+++ ChangeLog.md
@@ -6,6 +6,7 @@ be taken. This is a GA release and the changes described below indicate the chan
 BLOB
 * Fixed a bug that `DeleteRetentionPolicy.Days` should be `number` instead of `string` when calling `getServiceProperties`.
 * Added method `getAccountProperties` on `blobService`.
+* Added support for synchronous copy of block blobs and put block from URL.
 
 2018.05 Version 2.9.0-preview